Theblindpassenger.cpp
The seats of an Ultra Deluxe Semi-sleeper bus in Thuvax country are numbered according to the following rules:

Seat number 1 is reserved for the bus conductor.
Remaining seats are numbered as shown in the figure below.
The seating arrangement is as shown below.
RowNo Left   Right
      W  A   A  M  W
                  
      01 
1     02 03  04 05 06
2     11 10  09 08 07
3     12 13  14 15 16
4     21 20  19 18 17
5     22 ............    
   
A blind person enters the bus with a reservation ticket mentioning his seat number. He meets you at the start of the passage and shows you his ticket. Based on the seat number, help him find his seat by telling him the row number (1,2,...), seat position(window or aisle or middle), and the direction(left or right).